--- src/test/java/allure/jenkins/hm7/exercise/FailingTest.java
@@ -1,4 +1,0 @@
-package allure.jenkins.hm7.exercise;
-
-public class FailingTest {
-}
