--- src/test/java/allure/jenkins/hm7/exercise/ThreeTest.java
@@ -1,4 +1,0 @@
-package allure.jenkins.hm7.exercise;
-
-public class ThreeTest {
-}
